src/py-requirements.txt
python-socketio
requests
eventlet
